chore: rename AnimatedBlobs to Blobs for consistencies with Waves component
