+++ src/components/Blob.svelte
